backend/requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
python-jose[cryptography]
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-multipart
pydantic-settings
pydantic[email]
motor
pymongo
python-dotenv